+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.0
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
-**Date**:
+**Date**: Thu Jun 28 10:37:34 CEST 2018
 
 * API: [FIX] fixed set admin flag to false if admin flag was not send via PATCH /api/v2/users 
 * API: [CHANGE] added admin flag to /api/v2/users resource output
@@ -9,6 +9,7 @@
 * API: [FIX] fixed GET /api/v1/node/attributes and GET /api/v1/collection/children do now return the attribute "share"
 * API: [FIX] fixed GET /api/v1/collection/share now returns false if collection is not shared instead 422 - Balloon\Filesystem\Exception\Conflict
 * API: [FIX] fixed POST /api/v1/collection/share, acl is now correctly converted into v2 core
+* CORE: [FIX] fixed upgrade bug in Balloon\Migration\Delta\AddHashToHistory
 
 
 ## 2.0.0-rc1